backend/requirements.txt
Flask
Flask-CORS
Flask-SocketIO
gunicorn
python-dotenv
requests
opencv-python-headless
Pillow
librosa
soundfile
numpy
scipy
gevent
gevent-websocket
PyJWT
pymongo
cloudinary
scikit-learn
pandas
joblib
imagehash